Backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
firebase-admin>=6.4.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
google-generativeai>=0.3.0
apscheduler>=3.10.0